[Scheduler] `ratio(f, n, d = 1000, t = 0)`
Similar to `every()`, the `ratio()` method adds an event occuring `n` times within the duration `d`. The default is 1000 for it to be convenient to use with animations at a give frame rate, but this can also be set to a different value, such as 60000 for beats per minute instead.

When using this method, the callback function `f` will be called with four parameters, `(i, t, from, to)`, where `i` is the occurrence number. Together with `mod` this can make cycling through a sequence of frames easy; see the new example (phases of the moon) for a demonstration. Tests are added too.