[Tests] Find SVG icons and fix back button
In order to reuse tests from other repos, we look for the location of the test script and use the same prefix for loading the SVG.

The back button behaviour is also fixed by removing the iframe after the tests have run. This broke the script detection for missing tests, but this is replaced with a more generic timeout. Reporting is a bit improved to show missing tests as well.

Finally, a small drive-by change is made to utils.