before-update, ASAP
A new event notification, "before-update", is dispatched before updates begin (there can be several when there are time skips or rate changes).  This gives an opportunity to schedule events at the start of every update, or as soon as possible in the current update, using the `ASAP` symbol for the time in `at()` or period in `every()` (note that in the case of `every()`, the phase parameter is ignored). `ASAP` becomes the default for these values.

By definition, the concrete time for `ASAP` depends on the times at which updates happen but it is guaranteed to be at most the earliest time of other regularly scheduled events. It is still possible to use priority to order events when they happen at the same time.