Switch to end-exclusive intervals
Clock intervals are [from, to[, instead of ]from, to], which is more convenient as it removes the need for the Inclusive flag (which is thus removed entirely), or ASAP for at(), since we can just use from. The scheduler is also modified to only work forward and must be called with from < to; the clock can then process events backward when the rate is negative. Intervals of the form [n, d] are also improved.