[Scheduler] Delegate
The scheduler calls the delegate method `willScheduleEvents(from, to)` to let the delegate inject its own events in the schedule. Sorting will work as expected, ensuring that events happen in the right order based on time and priority (although no checks are made to the injected events as they may be out of range).

A helper function `at(f, t, priority?)` is provided to help the delegate create event objects with the right shape.